pyproject.toml
[project]
name = "docrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.45.0",
    "python-dotenv>=1.1.1",
    "httpx>=0.27.0",
    "google-genai>=0.3.0",
    "pymilvus>=2.6.0",
    "google>=3.0.0",
    "adk>=0.0.5",
    "google-adk>=1.11.0",
]
